Add files via upload
# Project Business Statistics: Axis Insurance
## BY: Kauda Ayodele Jimoh


## Define Problem Statement and Objectives

Leveraging customer information is of paramount importance for most businesses. In the case of an insurance company, the attributes of customers can be crucial in making business decisions. Hence, knowing to explore and generate value out of such data can be an invaluable skill to have.

Suppose you are hired as a Data Scientist in an Insurance company. The company wants to have a detailed understanding of the customer base for one of its Insurance Policy 'MediClaim'. The idea is to generate insights about the customers and answer a few key questions with statistical evidence, by using the past dataset. The dataset 'AxisInsurance' contains customers' details like age, sex, charges, etc. Perform the statistical analysis to answer the following questions using the collected data.

Explore the dataset and extract insights using Exploratory Data Analysis.

Prove(or disprove) that the medical claims made by the people who smoke are greater than those who don't?

Prove (or disprove) with statistical evidence that the BMI of females is different from that of males.

Does the smoking habit of customers depend on their region? [Hint: Create a contingency table using the pandas.crosstab() function]

Is the mean BMI of women with no children, one child, and two children the same? Explain your answer with statistical evidence.

The idea behind answering these questions is to help the company in making evidence-based business decisions.

## Assumptions

The Health Insurance Customer's data is a simple random sample from the population data, and the samples are independent of each other.

# Data Dictionary

AxisInsurance.csv contains the following information about customers of Axis Health Insurance.

  1. Age - This is an integer indicating the age of the primary beneficiary (excluding those above 64 years, since they are generally covered by the government).
  2. Sex - This is the policy holder's gender, either male or female.
  3. BMI - This is the body mass index (BMI), which provides a sense of how over or under-weight a person is relative to their height. BMI is equal to weight (in kilograms) divided by height (in meters) squared. An ideal BMI is within the range of 18.5 to 24.9.
  4. Children - This is an integer indicating the number of children / dependents covered by the insurance plan.
  5. Smoker - This is yes or no depending on whether the insured person regularly smokes tobacco.
  6. Region - This is the beneficiary's place of residence in the U.S., divided into four geographic regions - northeast, southeast, southwest, or northwest.
  7. Charges - Individual medical costs billed by health insurance